\name{psyo_rounds_map}

\alias{psyo_rounds_map}

\docType{data}

\title{
Map for data psyo_rounds and psyo_rounds2
}

\description{
Google map as ggplot object
}

\usage{data(psyo_rounds_map)}

\format{
See \code{\link[ggplot2]{ggplot}}
}

\examples{\donttest{
data(psyo_rounds_map)
plot(psyo_rounds_map)
}}

\keyword{datasets}
